retire container_internals.
